+++ Backend/APIBanca/Model/ViewModel/Cuenta/Cuenta_Empresa_DTO.cs
@@ -15,7 +15,7 @@ namespace Model.ViewModel.account
 
         // El resto en empresa
         public string cuit { get; set; }
-        public int account_id { get; set; }
+        public long account_id { get; set; }
         public string phone { get; set; }
         public string address { get; set; }
